[csc] Refactor to use client interface
In order for OperatorSources to directly create the datastore
CatalogSources, we will need to share code between both configuring
reconcilers. As such we need to ensure that both OperatorSource and
CatalogSourceConfig configuring reconcilers use identical types.

This commit modifies the CatalogSourceConfig Configuring Reconciler to
make use of the client interface. It also propagates the use of the
client interface to dependent methods such as the Purging Reconciler
and the registry interface.